client/requirements.txt
streamlit
requests
pillow
numpy
tenseal
torch
scipy
opencv-python
matplotlib
